Renaming of index app to feedback (#901)
renamed index app to feedback
