+++ app/feedback/app.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class FeedbackConfig(AppConfig):
+    name = "app.feedback"
