+++ app/feedback/factories/__init__.py
@@ -1,0 +1,2 @@
+from app.feedback.factories.bug_factory import BugFactory
+from app.feedback.factories.idea_factory import IdeaFactory
