+++ app/feedback/models/__init__.py
@@ -1,0 +1,3 @@
+from app.feedback.models.bug import Bug
+from app.feedback.models.feedback import Feedback
+from app.feedback.models.idea import Idea
