+++ app/feedback/models/bug.py
@@ -1,0 +1,5 @@
+from app.feedback.models.feedback import Feedback
+
+
+class Bug(Feedback):
+    pass
