+++ app/feedback/models/idea.py
@@ -1,0 +1,5 @@
+from app.feedback.models.feedback import Feedback
+
+
+class Idea(Feedback):
+    pass
