+++ app/feedback/serializers/__init__.py
@@ -1,0 +1,3 @@
+from app.feedback.serializers.bug import BugListSerializer
+from app.feedback.serializers.idea import IdeaListSerializer
+from app.feedback.serializers.feedback import FeedbackListPolymorphicSerializer
