+++ app/feedback/urls.py
@@ -1,7 +1,7 @@
 from django.urls import include, re_path
 from rest_framework import routers
 
-from app.index.views.feedback import FeedbackViewSet
+from app.feedback.views.feedback import FeedbackViewSet
 
 router = routers.DefaultRouter()
 router.register("feedbacks", FeedbackViewSet)
