--- app/index/app.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class IndexConfig(AppConfig):
-    name = "app.index"
