--- app/index/factories/__init__.py
@@ -1,2 +1,0 @@
-from app.index.factories.bug_factory import BugFactory
-from app.index.factories.idea_factory import IdeaFactory
