--- app/index/models/__init__.py
@@ -1,3 +1,0 @@
-from app.index.models.bug import Bug
-from app.index.models.feedback import Feedback
-from app.index.models.idea import Idea
