--- app/index/models/bug.py
@@ -1,5 +1,0 @@
-from app.index.models.feedback import Feedback
-
-
-class Bug(Feedback):
-    pass
