--- app/index/models/idea.py
@@ -1,5 +1,0 @@
-from app.index.models.feedback import Feedback
-
-
-class Idea(Feedback):
-    pass
