--- app/index/serializers/__init__.py
@@ -1,3 +1,0 @@
-from app.index.serializers.bug import BugListSerializer
-from app.index.serializers.idea import IdeaListSerializer
-from app.index.serializers.feedback import FeedbackListPolymorphicSerializer
