+++ app/tests/conftest.py
@@ -32,11 +32,11 @@ from app.emoji.factories.reaction_factory import (
     EventReactionFactory,
     NewsReactionFactory,
 )
+from app.feedback.factories import BugFactory, IdeaFactory
 from app.forms.tests.form_factories import FormFactory, SubmissionFactory
 from app.group.factories import GroupFactory, MembershipFactory
 from app.group.factories.fine_factory import FineFactory
 from app.group.factories.membership_factory import MembershipHistoryFactory
-from app.index.factories import BugFactory, IdeaFactory
 from app.kontres.factories import BookableItemFactory, ReservationFactory
 from app.payment.factories.order_factory import OrderFactory
 from app.payment.factories.paid_event_factory import PaidEventFactory
